Refactor: Extract OrderRow and OrderTable components in Admin page
