+++ Frontend/src/pages/Admin/Order/OrderRow.jsx
@@ -1,0 +1,5 @@
+export const OrderRow = () => {
+  return (
+    <div>OrderRow</div>
+  )
+}
