+++ Frontend/src/pages/Admin/Order/OrderTable.jsx
@@ -1,0 +1,7 @@
+import { OrderRow } from "./OrderRow"
+
+export const OrderTable = () => {
+  return (
+    <div>OrderTable</div>
+  )
+}
