+++ Frontend/src/pages/Admin/Product/ProductRow.jsx
